requirements.txt
Pillow==9.2.0
matplotlib==3.5.3